src/requirements.txt
fastapi
faiss-cpu
numpy
openai
pandas
python-dotenv
requests
sentence-transformers
streamlit
uvicorn[standard]